requirements.txt
Customtkinter
Pillow
